// src/shared/services/api/models/enrollmentDetailDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */
import type { RefTermDto } from './refTermDto';

export interface EnrollmentDetailDto {
  id?: string;
  user_id?: string;
  course_id?: string;
  enrolled_on?: string;
  /** @nullable */
  completed_on?: string | null;
  progress_percentage?: number;
  enrollment_status?: boolean;
  /** @nullable */
  course_title?: string | null;
  course_category?: RefTermDto;
  /** @nullable */
  user_full_name?: string | null;
}